refactor: email 로직 이동
