+++ backend/src/main/java/com/woowacourse/ternoko/support/application/EmailService.java
@@ -3,8 +3,8 @@ package com.woowacourse.ternoko.support.application;
 import com.woowacourse.ternoko.core.domain.interview.Interview;
 import com.woowacourse.ternoko.core.domain.interview.InterviewRepository;
 import com.woowacourse.ternoko.core.domain.interview.InterviewStatusType;
-import com.woowacourse.ternoko.support.alarm.EmailGenerator;
 import com.woowacourse.ternoko.support.application.response.EmailResponse;
+import com.woowacourse.ternoko.support.email.EmailGenerator;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
